refactor: Record 일부 기능 Records 클래스에 위임
